+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Basic detection for pseudo-elements in follow mode, such as switches using ::before as the only visible section
 - Right-click menu to websites (when in insert mode or when mouse is enabled)
 - Right-click menu for Vieb interface (when mouse is enabled)
+- Follow mode support for iframes
 
 ### Changed
 
@@ -27,6 +28,10 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Default order of elements in follow mode to align with functionality as it was before 3.3.0 (can still be shown with "action.reorderFollowLinks")
 - Only display a single line of the url if it's really long (overflow is suggested with ellipsis)
 - Pin command can now take an argument to search for a tab by name or index, similar to buffer commands
+
+### Fixed
+
+- Mouse events now get registered properly when clicking inside iframes
 
 ### Security